requirements.txt
Fabric==1.4.3
pycrypto==2.6
ssh==1.7.14
wsgiref==0.1.2
Jinja2==2.6
flask==0.9
Werkzeug==0.8.3
envoy==0.0.2
slimit==0.7.4
cssmin==0.1.4
boto==2.6.0
requests==1.1.0
python-tumblpy==0.7.0
git+https://github.com/jyr/python-tumblr.git@b91d98257
httplib2==0.7.7
oauth2==1.5.211
poster==0.8.1
